requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
redis==5.1.1

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.0
python-multipart==0.0.12

# Messaging (Twilio WhatsApp/SMS)
twilio==9.3.0
httpx==0.27.2

# Payments
stripe==10.12.0

# Task Queue
celery[redis]==5.4.0

# Utilities
python-dateutil==2.9.0
phonenumbers==8.13.47
